Revert "feat: add component Alert, MenuPrincipalAdm, NavBar, Search, SideNavBar. Change the functionality for the SideNavBar"
This reverts commit c6e3cbe7aa49f2683971297326634093608e5349.
