+++ .eslintrc.json
@@ -13,13 +13,5 @@
         , "vars": "local"
         , "ignoreRestSiblings": true
         , "caughtErrors": "none" }]
-    ,"react/no-unknown-property": ["error",{
-           "ignoreRestSiblings": true
-            , "caughtErrors": "none"
-            , "ignore": [
-              "jsx"
-            ]
-          }
-        ]
   }
 }
