--- 3.3.1
@@ -1,38 +1,0 @@
-Progress: resolved 0, reused 1, downloaded 0, added 0
-Progress: resolved 13, reused 12, downloaded 0, added 0
-Progress: resolved 24, reused 23, downloaded 0, added 0
-Progress: resolved 24, reused 24, downloaded 0, added 0
-Progress: resolved 27, reused 26, downloaded 0, added 0
-Progress: resolved 29, reused 28, downloaded 0, added 0
-Progress: resolved 33, reused 32, downloaded 0, added 0
-Progress: resolved 34, reused 32, downloaded 0, added 0
-Progress: resolved 50, reused 35, downloaded 0, added 0
-Progress: resolved 65, reused 49, downloaded 0, added 0
-Progress: resolved 78, reused 62, downloaded 0, added 0
-Progress: resolved 89, reused 72, downloaded 0, added 0
-Progress: resolved 117, reused 100, downloaded 0, added 0
-Progress: resolved 142, reused 125, downloaded 0, added 0
-Progress: resolved 207, reused 190, downloaded 0, added 0
-Progress: resolved 225, reused 208, downloaded 0, added 0
-Progress: resolved 229, reused 212, downloaded 0, added 0
-Progress: resolved 240, reused 223, downloaded 0, added 0
-Progress: resolved 267, reused 252, downloaded 0, added 0
-Progress: resolved 374, reused 359, downloaded 0, added 0
-Progress: resolved 433, reused 418, downloaded 1, added 0
-Progress: resolved 468, reused 453, downloaded 1, added 0
-Progress: resolved 484, reused 469, downloaded 1, added 0
-Progress: resolved 498, reused 483, downloaded 1, added 0
-Progress: resolved 550, reused 535, downloaded 1, added 0
-Progress: resolved 582, reused 567, downloaded 1, added 0
-Progress: resolved 603, reused 588, downloaded 1, added 0
-Packages: +5 -4
-+++++----
-Progress: resolved 603, reused 588, downloaded 1, added 1
-Progress: resolved 603, reused 588, downloaded 1, added 4
-Progress: resolved 603, reused 588, downloaded 2, added 5, done
-
-dependencies:
-+ typescript 5.0.3
-
-The integrity of 1862 files was checked. This might have caused installation to take longer.
-Done in 48.4s
