+++ package.json
@@ -1,52 +1,45 @@
 {
-  "name": "icalidad23q2",
-  "version": "0.1.3",
+  "name": "icalidad22q1",
+  "version": "0.1.0",
   "private": true,
   "scripts": {
     "dev": "next dev ",
-    "build": "next build ",
+    "build": "next build",
     "start": "next start",
     "lint": "next lint"
   },
   "dependencies": {
-    "@chakra-ui/react": "^2.5.1",
-    "@emotion/react": "^11.10.6",
-    "@emotion/styled": "^11.10.6",
-    "@headlessui/react": "^1.7.13",
-    "@heroicons/react": "^2.0.16",
-    "@prisma/client": "^4.11.0",
-    "@tailwindcss/forms": "^0.5.3",
-    "autoprefixer": "^10.4.14",
-    "bcrypt": "^5.1.0",
+    "@headlessui/react": "^1.7.5",
+    "@heroicons/react": "^1.0.6",
+    "@prisma/client": "^4.3.1",
+    "bcrypt": "^5.0.1",
     "cors": "^2.8.5",
-    "daisyui": "^2.51.4",
     "eslint": "^7.23.0",
-    "eslint-import-resolver-typescript": "^3.5.3",
-    "eslint-plugin-import": "^2.27.5",
+    "eslint-import-resolver-typescript": "^3.5.2",
+    "eslint-plugin-import": "^2.24.1",
     "formik": "^2.2.9",
-    "framer-motion": "^10.6.0",
-    "mssql": "^9.1.1",
+    "framer-motion": "^8.5.0",
+    "mssql": "^7.2.0",
     "next": "^12.1.6",
-    "next-auth": "^4.20.1",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-hook-form": "^7.43.6",
-    "react-icons": "^4.8.0",
-    "sharp": "^0.32.0",
-    "swr": "^2.1.0",
-    "tedious": "^15.1.3",
-    "tsconfig-paths": "^4.1.2",
-    "tw-elements": "^1.0.0-beta1",
-    "typescript": "^5.0.3",
-    "yup": "^1.0.2"
+    "next-auth": "^4.12.1",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-hook-form": "^7.32.1",
+    "react-icons": "^4.7.1",
+    "swr": "^1.3.0",
+    "tedious": "^11.0.8",
+    "tsconfig-paths": "^4.1.1",
+    "tw-elements": "^1.0.0-alpha13"
   },
   "devDependencies": {
-    "@types/node": "18.15.3",
-    "@types/react": "18.0.28",
+    "@tailwindcss/forms": "^0.5.3",
+    "@types/node": "18.11.18",
+    "@types/react": "18.0.26",
+    "autoprefixer": "^10.4.7",
     "eslint-config-next": "^12.0.4",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "postcss": "^8.4.21",
-    "prisma": "^4.11.0",
-    "tailwindcss": "^3.3.1"
+    "postcss": "^8.4.13",
+    "prisma": "^4.3.1",
+    "tailwindcss": "^3.2.4"
   }
 }
