+++ pages/AdminAction/index.jsx
@@ -1,15 +1,9 @@
-//import HomeDashBoard from "../DashBoard";
-import Prueba from "../DashBoard/Prueba";
+import React from 'react'
 
-const AdminAction = () => {
-
+const index = () => {
   return (
-    <>
-      <Prueba>
-        <div className="text-4xl h-full text-black text-center  bg-white ">Admin Action</div>
-      </Prueba>
-    </>
-  );
+    <div>index</div>
+  )
 }
 
-export default AdminAction
+export default index
