+++ pages/AdminAudit/index.jsx
@@ -1,16 +1,9 @@
+import React from 'react'
 
-//import HomeDashBoard from "../DashBoard";
-import Prueba from "../DashBoard/Prueba";
-
-const AdminAudit = () => {
-
+const index = () => {
   return (
-    <>
-      <Prueba>
-        <div className="text-4xl h-full text-black text-center  bg-white ">Admin Audit</div>
-      </Prueba>
-    </>
-  );
+    <div>index</div>
+  )
 }
 
-export default AdminAudit
+export default index
