+++ pages/AdminDocument/index.jsx
@@ -1,14 +1,14 @@
-
-//import HomeDashBoard from "../DashBoard";
-import Prueba from "../DashBoard/Prueba";
+import { useContext } from "react";
+import { IcalidadContext } from "../Contexts/IcalidadContext";
+import HomeDashBoard from "../DashBoard";
 
 const AdminDocument = () => {
-
+  const { menuDataMenu, setDataMenu } = useContext(IcalidadContext);
   return (
     <>
-      <Prueba>
-        <div className="text-4xl h-full text-black text-center  bg-white ">Admin Document</div>
-      </Prueba>
+      <HomeDashBoard>
+        <div className="text-4xl  bg-slate-500 items-center text-slate-500">Admin iCalidad</div>
+      </HomeDashBoard>
     </>
   );
 }
