+++ pages/Components/index.jsx
@@ -1,12 +1,3 @@
-//export { default as SideNavBar } from '../DashBoard/SideNavBar' ;
-//export { default as NavBar } from './NavBar' ;
-
-import React from 'react'
-
-const index = () => {
-  return (
-    <div>index</div>
-  )
-}
-
-export default index
+export { default as LinkC } from './LinkC' ;
+export { default as SideNavBar } from './SideNavBar' ;
+export { default as NavBar } from './NavBar' ;
