+++ pages/DashBoard/Avatar.jsx
@@ -1,22 +1,9 @@
-import Image from 'next/image'
-import avatarImage from '../../public/photo-Avatar1.jpg'
+import React from 'react'
 
-const Avatar = ({ userImage }) => {
+const avatar = () => {
   return (
-    <>
-   <div className="avatar online relative">
-    <div className="w-12 rounded-full ring ring-primary ring-offset-base-100 ring-offset-1">
-      <Image src={avatarImage}  alt='Avatar' className='rounded-full' layout='fill' ></Image>
-    </div>
-  </div>
-  </>
+    <div>avatar</div>
   )
 }
 
-export default Avatar
-
-// .color1 { #34acec };
-// .color2 { #99d4f4 };
-// .color3 { #0b5379 };
-// .color4 { #7aa9c4 };
-// .color5 { #acacbc };
+export default avatar
