--- pages/DashBoard/Search.jsx
@@ -1,12 +1,0 @@
-
-const Search = () => {
-  return (
-    <>
-         <div className="form-control">
-            <input type="text" placeholder="Search" className="input input-bordered" />
-        </div>
-    </>
-  )
-}
-
-export default Search
