+++ pages/api/User/[...dashBoardUser].jsx
@@ -1,8 +1,6 @@
 import { usegetPool, typeParameter } from "lib/database/connection";
 
 export default async function handlerUserAction(req, res) {
-  //console.log("req", req.method);
-  
   const requestOptions = {
     headers: { "Content-Type": "application/json" },
     body: JSON.stringify(req.body),
@@ -14,18 +12,16 @@ export default async function handlerUserAction(req, res) {
         const typeParam = await typeParameter();
         const pool = await usegetPool("Default");
         const request = await pool.request();
-        //console.log('dashBoardUser', dashBoardUser);
         await request.input("p_IdEmpleado", typeParam.Int, dashBoardUser[1]);
         await request.input("p_Version", typeParam.Int, dashBoardUser[2]);
         await request.input("p_IdMenuFather", typeParam.Int, dashBoardUser[3]);
         const result = await request.execute("PF_Gen_TMenu");
-        const { rowsAffected , recordsets } = result;
-        //console.log(JSON.stringify(recordsets));
-        //const recordset = result.recordset;
-        
-        res.status(200).json({ recordsets, rowsAffected });
+        //console.log(JSON.stringify(result.result));
+        const recordset = result.recordset;
+        //const { rowsAffected , recordsets } = result;
+        res.status(200).json({ recordset });
       } catch (error) {
-        console.log('mensaje de error: ', error);
+        console.log(error);
         res.status(500).json({ error: error.message });
       }
