+++ pages/contexts/IcalidadContext.jsx
@@ -17,5 +17,3 @@ export const IcalidadProvider = ({ children }) => {
     </IcalidadContext.Provider>
   )
 }
-
-export default IcalidadProvider;
