+++ styles/globals.css
@@ -31,7 +31,9 @@ a {
   text-decoration: none;
 }
 
-
+* {
+  box-sizing: border-box;
+}
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
